+++ cognite/client/_api/data_modeling/containers.py
@@ -10,7 +10,7 @@ from cognite.client.data_classes.data_modeling.containers import (
     ContainerFilter,
     ContainerList,
 )
-from cognite.client.data_classes.data_modeling.ids import ContainerIdentifier, DataModelingId, load_identifier
+from cognite.client.data_classes.data_modeling.ids import ContainerIdentifier, DataModelingId, _load_identifier
 
 
 class ContainersAPI(APIClient):
@@ -85,7 +85,7 @@ class ContainersAPI(APIClient):
         ...
 
     def retrieve(self, ids: ContainerIdentifier | Sequence[ContainerIdentifier]) -> Container | ContainerList | None:
-        """`Retrieve one or more container by id(s). <https://docs.cognite.com/api/v1/#tag/Containers/operation/byExternalIdsContainers>`_
+        """`Retrieve one or more container by id(s). <https://docs.cognite.com/api/v1/#operation/byExternalIdsContainers>`_
 
         Args:
             ids (ContainerId | Sequence[ContainerId]): Identifier for container(s).
@@ -106,11 +106,11 @@ class ContainersAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> res = c.data_modeling.containers.retrieve(DataModelingId(space='mySpace', external_id='myContainer'))
         """
-        identifier = load_identifier(ids, "container")
+        identifier = _load_identifier(ids, "container")
         return self._retrieve_multiple(list_cls=ContainerList, resource_cls=Container, identifiers=identifier)
 
     def delete(self, id: ContainerIdentifier | Sequence[ContainerIdentifier]) -> list[DataModelingId]:
-        """`Delete one or more containers <https://docs.cognite.com/api/v1/#tag/Containers/operation/deleteContainers>`_
+        """`Delete one or more containers <https://docs.cognite.com/api/v1/#operation/deleteContainers>`_
 
         Args:
             id (ContainerId | Sequence[ContainerId): The container identifier(s).
@@ -126,14 +126,14 @@ class ContainersAPI(APIClient):
         """
         deleted_containers = cast(
             list,
-            self._delete_multiple(identifiers=load_identifier(id, "container"), wrap_ids=True, returns_items=True),
+            self._delete_multiple(identifiers=_load_identifier(id, "container"), wrap_ids=True, returns_items=True),
         )
         return [DataModelingId(space=item["space"], external_id=item["externalId"]) for item in deleted_containers]
 
     def list(
         self, space: str | None = None, limit: int = DATA_MODELING_LIST_LIMIT_DEFAULT, include_global: bool = False
     ) -> ContainerList:
-        """`List containers <https://docs.cognite.com/api/v1/#tag/Containers/operation/listContainers>`_
+        """`List containers <https://docs.cognite.com/api/v1/#operation/listContainers>`_
 
         Args:
             space (int, optional): The space to query
@@ -185,7 +185,7 @@ class ContainersAPI(APIClient):
         ...
 
     def apply(self, container: ContainerApply | Sequence[ContainerApply]) -> Container | ContainerList:
-        """`Add or update (upsert) containers. <https://docs.cognite.com/api/v1/#tag/Containers/operation/ApplyContainers>`_
+        """`Add or update (upsert) containers. <https://docs.cognite.com/api/v1/#operation/ApplyContainers>`_
 
         Args:
             container (container: ContainerApply | Sequence[ContainerApply]): Container or containers of containers to create or update.