+++ scripts/create_data_models.py
@@ -6,9 +6,8 @@ from pathlib import Path
 
 import yaml
 
+import cognite.client.data_classes.data_modeling as m
 from cognite.client import CogniteClient
-from cognite.client.data_classes.data_modeling import Space
-from cognite.client.data_classes.data_modeling.data_models import DataModelApply
 
 
 def dump(client: CogniteClient):
@@ -26,37 +25,58 @@ def dump(client: CogniteClient):
 
 
 def copy_pygen_test_data(pygen: CogniteClient, client: CogniteClient):
-    sdk_integration = Space(
-        space="sdkIntegrationTests",
-        description="Space used for integration testing in the SDK",
-        name="SDK Integration Testing",
-    )
-    immutable_space = Space(
+    # sdk_integration = Space(
+    #     space="sdkIntegrationTests",
+    #     description="Space used for integration testing in the SDK",
+    #     name="SDK Integration Testing",
+    # )
+    space = m.SpaceApply(
         space="IntegrationTestsImmutable",
         description="Space used for integration testing in the SDK",
         name="SDK Integration Testing copy from Pygen",
     )
-    empty_data_model = DataModelApply(
-        space=sdk_integration.space, external_id="integrationTestEmptyModel", version="v0"
+    # empty_data_model = DataModel(space=sdk_integration.space, external_id="integrationTestEmptyModel", version="v0")
+    # client.data_modeling.data_models.apply(empty_data_model)
+    # print("Empty data model added")
+    #
+    # _ = client.data_modeling.spaces.apply(immutable_space)
+    # print("Space added")
+    #
+    # containers = pygen.data_modeling.containers.list(-1)
+    # client.data_modeling.containers.apply(containers)
+    # print("Containers added")
+    #
+    client.data_modeling.views.list(-1)
+    # client.data_modeling.views.apply(views)
+    #
+    # print("Views added")
+    #
+    # data_models = pygen.data_modeling.data_models.list(-1, space=immutable_space.space)
+    # client.data_modeling.data_models.apply(data_models)
+    # print("Data Models added")
+    # edges = pygen.data_modeling.instances.list(instance_type="edge", limit=-1)
+    # print(edges)
+
+    # for view in views:
+    #     source = view.as_reference()
+    #
+    #     nodes = pygen.data_modeling.instances.list(instance_type="node", limit=-1, sources=[source])
+    #     apply_nodes = [n.as_apply(source, 1) for n in nodes]
+    #     created_nodes = client.data_modeling.instances.apply(apply_nodes)
+    #     print(source)
+    #     print(created_nodes)
+    actor_view = pygen.data_modeling.views.retrieve((space.space, "Person"))
+    ingested_edges = pygen.data_modeling.instances.list(
+        instance_type="node", limit=-1, sources=actor_view[0].as_reference()
     )
-    client.data_modeling.data_models.apply(empty_data_model)
-    print("Empty data model added")
-
-    _ = client.data_modeling.spaces.apply(immutable_space)
-    print("Space added")
-
-    containers = pygen.data_modeling.containers.list(-1)
-    client.data_modeling.containers.apply(containers)
-    print("Containers added")
-
-    views = pygen.data_modeling.views.list(-1)
-    client.data_modeling.views.apply(views)
-
-    print("Views added")
-
-    data_models = pygen.data_modeling.data_models.list(-1, space=immutable_space.space)
-    client.data_modeling.data_models.apply(data_models)
-    print("Data Models added")
+    if len(ingested_edges) > 0:
+        print("Skipping edges")
+        return
+    # for view in views:
+    edges = pygen.data_modeling.instances.list(instance_type="edge", limit=-1)
+    apply_edges = [n.as_apply(None, None) for n in edges]
+    created_edges = client.data_modeling.instances.apply(apply_edges)
+    print(created_edges)
 
 
 if __name__ == "__main__":