+++ CHANGELOG.md
@@ -17,6 +17,27 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [6.4.3] - 2023-06-15
+### Added
+- Accept direct relation values as tuples in `EdgeApply`
+
+## [6.4.2] - 2023-06-15
+### Changed
+- When providing ids as tuples in `instances.retrieve` and `instances.delete` you should not 
+have to specify the instance type in each tuple
+
+### Fixed
+- Bug where edges and nodes would get mixed up on `instances.retrieve`
+
+## [6.4.1] - 2023-06-14
+### Fixed
+- Add the missing page_count field for diagram detect items.
+
+## [6.4.0] - 2023-06-12
+### Added
+- Partial support for the instance resource in the Data Modeling API with the implementation 
+  `client.data_modeling.instances`, the endpoints `list`, `delete`, `retrieve`, and `apply`
+
 ## [6.3.2] - 2023-06-08
 ### Fixed
 - Requests being retried around a token refresh cycle, no longer risk getting stuck with an outdated token.